Factor grouping logic into separate file
